% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecorisk-package.R
\docType{package}
\name{ecorisk-package}
\alias{ecorisk}
\alias{ecorisk-package}
\title{ecorisk: Risk Assessments for Ecosystems or Ecosystem Components}
\description{
Implementation of a modular framework for ecosystem risk assessments, combining existing risk assessment approaches tailored to semi-quantitative and quantitative analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HeleneGutte/ecorisk}
  \item Report bugs at \url{https://github.com/HeleneGutte/ecorisk/issues}
}

}
\author{
\strong{Maintainer}: Helene Gutte \email{helenegutte96@gmail.com} (\href{https://orcid.org/0009-0006-0681-0444}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Saskia A. Otto (\href{https://orcid.org/0000-0001-7780-1322}{ORCID})
}

}
\keyword{internal}
