% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_expert_status}
\alias{ex_expert_status}
\title{Expert-based status scores for four indicators}
\format{
A data frame with 4 observations and 3 variables.
\describe{
\item{indicator}{Name of the assessed indicator.}
\item{status}{Current status of each indicator, either 'good' or 'undesired'.}
\item{score}{Score for each status (+1 or -1), will be combined with the
vulnerability to derive risk.}
}
}
\usage{
ex_expert_status
}
\description{
This demo dataset contains the status scores of four Baltic Sea indicators based
on expert knowledge. The format is the same as the output table from the
\code{\link{status}} function that evaluates the status based on time series.
}
\keyword{datasets}
