% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_aggregate_risk_expert}
\alias{ex_output_aggregate_risk_expert}
\title{Example output from the \code{aggregate_risk()} function based on expert scores}
\format{
A list of three data frames.
\describe{
\item{multi_indicator_risk}{A data frame with 30 rows and 5 columns, containing the
multi-indicator risk and uncertainty of each pressure per type and pathway.}
\item{multi_pressure_risk}{A data frame with 24 rows and 5 columns, containing the
multi-pressure risk and uncertainty on each indicator per type and pathway.}
\item{ecosystem_risk}{A data frame with 6 rows and 4 columns, containing the
aggregated ecosystem risk and uncertainty per type and pathway.}
}
}
\usage{
ex_output_aggregate_risk_expert
}
\description{
This is an expert-based example output from the \code{\link{aggregate_risk}} function
applied to the \code{\link{ex_output_risk_expert}} demo data.
}
\keyword{datasets}
