% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_aggregate_risk_model}
\alias{ex_output_aggregate_risk_model}
\title{Example output from the \code{aggregate_risk()} function based on modelled scores}
\format{
A list of three data frames.
\describe{
\item{multi_indicator_risk}{A data frame with 32 rows and 5 variables,
containing the multi-indicator risk and uncertainty of each pressure per
type and pathway.}
\item{multi_pressure_risk}{A data frame with 8 rows and 5 variables, containing
the multi-pressure risk and uncertainty on each indicator per type and
pathway.}
\item{ecosystem_risk}{A data frame with 4 rows and 4 variables, containing
the aggregated ecosystem risk and uncertainty per type and pathway.}
}
}
\usage{
ex_output_aggregate_risk_model
}
\description{
This dataset provides example output from the \code{\link{aggregate_risk}} function,
applied to the \code{\link{ex_output_risk_model}} demo data, following the modelling
pathway.
}
\keyword{datasets}
