% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_calc_exposure}
\alias{ex_output_calc_exposure}
\title{Example output from the \code{calc_exposure()} function}
\format{
A data frame with 5 observations and 3 variables.
\describe{
\item{pressure}{Name of the assessed pressure.}
\item{exposure}{Calculated combined score of all exposure components.}
\item{uncertainty}{Calculated combined score of associated uncertainties.}
}
}
\usage{
ex_output_calc_exposure
}
\description{
This dataset provides an expert-based example output from the
\code{\link{calc_exposure}} function applied to \code{\link{ex_expert_exposure}}
demo data.
}
\keyword{datasets}
