% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_calc_sensitivity}
\alias{ex_output_calc_sensitivity}
\title{Example output from the \code{calc_sensitivity()} function}
\format{
A data frame with 40 observations and 18 variables.
\describe{
\item{indicator}{Name of the assessed indicator.}
\item{pressure}{Name of the assessed pressure.}
\item{type}{Effect type (direct, indirect, or direct + indirect).}
\item{pathway}{Pathway with which sensitivity has been assessed (expert- or model-based).}
\item{sensitivity}{Combined sensitivity score.}
\item{adaptive_capacity}{Combined adaptive capacity score.}
\item{uncertainty_sens}{Combined score of the associated sensitivity uncertainties.}
\item{uncertainty_ac}{Combined score of the associated adaptive capacity uncertainties.}
\item{sens_original.sens_feeding}{Original sensitivity score for the feeding trait.}
\item{sens_original.sens_behaviour}{Original sensitivity score for the behaviour trait.}
\item{sens_original.sens_reproduction}{Original sensitivity score for the reproduction trait.}
\item{sens_original.sens_habitat}{Original sensitivity score for the habitat trait.}
\item{sens_original.sens_general}{Original general sensitivity score.}
\item{ac_original.ac_feeding}{Original adaptive capacity score for the feeding trait.}
\item{ac_original.ac_behaviour}{Original adaptive capacity score for the behaviour trait.}
\item{ac_original.ac_reproduction}{Original adaptive capacity score for the reproduction trait.}
\item{ac_original.ac_habitat}{Original adaptive capacity score for the habitat trait.}
\item{ac_original.ac_general}{Original general adaptive capacity score.}
}
}
\usage{
ex_output_calc_sensitivity
}
\description{
This is an expert-based example output from the \code{\link{calc_sensitivity}} function
applied to the \code{\link{ex_expert_sensitivity}} demo data.
}
\keyword{datasets}
