% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_risk_expert}
\alias{ex_output_risk_expert}
\title{Example output from the \code{risk()} function based on expert scores}
\format{
A data frame with 40 observations and 8 variables.
\describe{
\item{indicator}{Name of the assessed indicator.}
\item{pressure}{Name of the assessed pressure.}
\item{type}{Effect type (direct, indirect, or direct + indirect).}
\item{pathway}{Pathway used for the exposure and sensitivity assessment.}
\item{vulnerability}{Vulnerability score.}
\item{status}{Qualitative descriptor of the current status of the indicator.}
\item{risk}{Risk score.}
\item{uncertainty}{Uncertainty score, associated with the vulnerability component scoring.}
}
}
\usage{
ex_output_risk_expert
}
\description{
This is an expert-based example output from the \code{\link{risk}} function
applied to the \code{\link{ex_output_vulnerability_expert}} and \code{\link{ex_expert_status}}
demo datasets.
}
\keyword{datasets}
