% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_risk_model}
\alias{ex_output_risk_model}
\title{Example output from the \code{aggregate_risk()} function based on modelled scores}
\format{
A data frame with 16 observations and 8 variables.
\describe{
\item{indicator}{Name of the assessed indicator.}
\item{pressure}{Name of the assessed pressure.}
\item{type}{Type of effect (always direct + indirect for modelling pathway).}
\item{pathway}{Pathway used for the exposure and sensitivity assessment.}
\item{vulnerability}{Vulnerability score.}
\item{status}{Qualitative descriptor of the current status of the indicator.}
\item{risk}{Risk score.}
\item{uncertainty}{Uncertainty score associated with the vulnerability component
scoring.}
}
}
\usage{
ex_output_risk_model
}
\description{
This dataset provides example output from the \code{\link{risk}} function,
applied to the \code{\link{ex_output_vulnerability_model}} and \code{\link{ex_output_status}}
demo datasets, following the modelling pathway.
}
\keyword{datasets}
