% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_status}
\alias{ex_output_status}
\title{Example output from the \code{status()} function}
\format{
A data frame with 2 rows and 3 variables.
\describe{
\item{indicator}{Name of the assessed indicator.}
\item{status}{Qualitative description of the current status compared to the threshold.}
\item{score}{Score used in the risk function to calculate risk from vulnerability and status.}
}
}
\usage{
ex_output_status
}
\description{
This dataset provides example output from the \code{\link{status}} function,
applied to four Baltic Sea indicator time series provided in \code{\link{indicator_ts_baltic}}.
}
\keyword{datasets}
