% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_vulnerability_expert}
\alias{ex_output_vulnerability_expert}
\title{Example output from the \code{vulnerability()} function based on expert scores}
\format{
A data frame with 40 observations and 6 variables.
\describe{
\item{indicator}{Name of the assessed indicator.}
\item{pressure}{Name of the assessed pressure.}
\item{type}{Effect type (direct, indirect, or direct + indirect).}
\item{pathway}{Pathway used for the exposure and sensitivity assessment.}
\item{vulnerability}{Vulnerability score.}
\item{uncertainty}{Uncertainty associated with the vulnerability score.}
}
}
\usage{
ex_output_vulnerability_expert
}
\description{
This dataset contains an expert-based example output from the
\code{\link{vulnerability}} function applied to \code{\link{ex_output_calc_exposure}}
and \code{\link{ex_output_calc_sensitivity}} demo datasets.
}
\keyword{datasets}
