% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_output_vulnerability_model}
\alias{ex_output_vulnerability_model}
\title{Example output from the \code{vulnerability()} function based on modelled scores}
\format{
A data frame with 16 observations and 6 variables.
\describe{
\item{indicator}{Names of the assessed indicator.}
\item{pressure}{Names of the assessed pressure.}
\item{type}{Type of effect (always direct + indirect for modelling pathway).}
\item{pathway}{Pathway used for the exposure and sensitivity assessment.}
\item{vulnerability}{Vulnerability score for each pressure-indicator-type combination.}
\item{uncertainty}{Uncertainty associated with the vulnerability score.}
}
}
\usage{
ex_output_vulnerability_model
}
\description{
This dataset provides example output from the \code{\link{vulnerability}} function,
applied to the \code{\link{ex_output_model_exposure}} and \code{\link{ex_output_model_sensitivity}}
demo datasets, following the modelling pathway.
}
\keyword{datasets}
