% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_ts_baltic}
\alias{indicator_ts_baltic}
\title{Baltic Sea indicator time series}
\format{
A data frame with 33 observations and 3 variables.
\describe{
\item{year}{Time variable.}
\item{zooplankton_mean_size}{Mean size of zooplankton (in wet weight micrograms).}
\item{eastern_baltic_cod}{Mean spawning stock biomass of eastern Baltic cod (units unspecified).}
}
}
\usage{
indicator_ts_baltic
}
\description{
Time series of two marine indicators covering the period 1984–2016 in the
Eastern Baltic Sea (data altered from original time series). This dataset
serves as a demo input in the \code{\link{model_sensitivity}} and
\code{\link{model_exposure}} functions.
}
\keyword{datasets}
