% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pressure_ts_northsea}
\alias{pressure_ts_northsea}
\title{North Sea pressure time series}
\format{
A data frame with 33 observations and 9 variables.
\describe{
\item{year}{Time variable.}
\item{bot_temp}{Mean sea bottom temperature in the Baltic Sea per year (in °C).}
\item{bot_sal}{Mean sea bottom salinity in the Baltic Sea per year.}
\item{fishing_cod}{Mean eastern Baltic cod fishing pressure per year.}
}
}
\usage{
pressure_ts_northsea
}
\description{
Time series of three environmental and anthropogenic pressures in the North
Sea covering the period 1970–2020 (data altered from original time series).
This dataset serves as internal test data.
}
\keyword{datasets}
